Refactor Code to Match Project Style
